refactor(models): split forum reads and web handlers further
Continue decomposition by moving listing/crud/moderation handlers into focused files, extracting shared request helpers, and splitting forum model read/list/write/user helper concerns while preserving behavior and routes.

Made-with: Cursor
