+++ internal/models/forum_helpers.go
@@ -1,0 +1,12 @@
+package models
+
+func containsAny(source []string, target []string) bool {
+	for _, s := range source {
+		for _, t := range target {
+			if s == t {
+				return true
+			}
+		}
+	}
+	return false
+}
